+++ terratex_reallife/SQLS/update.1.1.0.sql
@@ -1,8 +1,10 @@
 CREATE TABLE `user_gold` (
   `Nickname` varchar(255) NOT NULL,
   `Gold` int(11) DEFAULT '0',
-  PRIMARY KEY (`Nickname`),
-  CONSTRAINT `user_gold_user_Nickname_fk` FOREIGN KEY (`Nickname`) REFERENCES `user` (`Nickname`) ON DELETE CASCADE ON UPDATE CASCADE
+  `FoodBooster` int(11) DEFAULT '0',
+  `HufeisenBooster` int(11) DEFAULT '0',
+  `FuelBooster` int(11) DEFAULT '0',
+  PRIMARY KEY (`Nickname`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 ALTER TABLE user_licenses ADD industrialFishingLic INT(1) DEFAULT 0 NULL;